Adding project

+++ Library Management System/Remove_book.py
@@ -8,7 +8,14 @@ def Remove_book():
           "Removing a Book Data-\n"
           "*********************")
     book_id = input("Enter the book ID ( 4-Digit ): ").strip()
-    delete = f'delete from Library where BookID = "{book_id}"'
-    cur.execute(delete)
-    con.commit()
-    print('\n The Book Data Successfully Deleted!')
+    sql = f"select * from Library where BookID = '{book_id}'"
+    cur.execute(sql)
+    res = cur.fetchall()
+    len1 = len(res)
+    if len1 == 1:
+        delete = f'delete from Library where BookID = "{book_id}"'
+        cur.execute(delete)
+        con.commit()
+        print('\nThe Book Data Successfully Deleted!\n')
+    else:
+        print(f"\nBook With {book_id} this ID is not in the Library!\n")
